Shops testing

+++ controllers/sellerController.js
@@ -94,7 +94,6 @@ class SellerController {
       );
       res.status(200).json({ message: "Success update seller" });
     } catch (error) {
-      console.log(error);
       next(error);
     }
   }
@@ -102,6 +101,10 @@ class SellerController {
   static async delete(req, res, next) {
     try {
       const { id } = req.params;
+      const seller = await Seller.findByPk(id);
+      if (!seller) {
+        throw { name: "not_found" };
+      }      
       await Seller.destroy({
         where: {
           id,
